server_py/requirements.txt
grpcio>=1.66.2
grpcio-tools>=1.66.2
protobuf>=5.28.0
nltk>=3.9.1
transformers>=4.49.0
torch>=2.6.0